move launcher back
