+++ packages/client/hmi-client/src/components/models/forecast-launcher.vue
